\name{LOSim-package}
\alias{LOSim-package}
\alias{LOSim}
\docType{package}
\title{
\packageTitle{LOSim}
}
\description{
\packageDescription{LOSim}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LOSim}
\packageIndices{LOSim}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{LOSim}

Maintainer: \packageMaintainer{LOSim}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
